app/requirements.txt
# requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.1
gunicorn==20.1.0
Werkzeug>=3.0.0